refactor(api): move connections related structs to its own mod
